Change Progress datatype to double and uploading glitch DONE also the rounding off of the progress when retrieving and vieweing it on the table.

+++ ELearningV1/Models/ViewModel/VMViewEmployeeCourseStatus.cs
@@ -10,7 +10,7 @@ namespace ELearningV1.Models.ViewModel
         public string EmployeeNumber { get; set; }
         public string EmpName { get; set; }
         public string Course { get; set; }
-        public int Progress { get; set; }
+        public double Progress { get; set; }
         public float? Score { get; set; }
         public string Status1 { get; set; }
         public DateTime EnrolledDate { get; set; }
